+++ src/ui/first_run/main.rs
@@ -135,7 +135,7 @@ impl SimpleComponent for FirstRunApp {
                 .forward(sender.input_sender(), std::convert::identity),
 
             default_paths: DefaultPathsApp::builder()
-                .launch(false)
+                .launch(())
                 .forward(sender.input_sender(), std::convert::identity),
 
             download_components: DownloadComponentsApp::builder()